Move Azure specific functions to AzureResourceValidationUtils
