+++ service/src/main/java/bio/terra/workspace/service/resource/ResourceValidationUtils.java
@@ -13,7 +13,6 @@ import bio.terra.workspace.common.utils.GcpUtils;
 import bio.terra.workspace.common.utils.Rethrow;
 import bio.terra.workspace.db.ResourceDao;
 import bio.terra.workspace.db.exception.FieldSizeExceededException;
-import bio.terra.workspace.generated.model.ApiAzureVmCreationParameters;
 import bio.terra.workspace.generated.model.ApiGcpAiNotebookInstanceCreationParameters;
 import bio.terra.workspace.generated.model.ApiGcpAiNotebookInstanceVmImage;
 import bio.terra.workspace.service.policy.TpsApiDispatch;
@@ -73,23 +72,6 @@ public class ResourceValidationUtils {
       "Invalid GCS bucket name specified. Names must be 3-222 lowercase letters, numbers, and dashes. See Google documentation for the full specification.";
 
   /**
-   * Azure Storage Account name validation valid. An storage account name must be between 3-24
-   * characters in length and may contain numbers and lowercase letters only.
-   */
-  public static final Pattern AZURE_STORAGE_ACCOUNT_NAME_VALIDATION_PATTERN =
-      Pattern.compile("^[a-z0-9]{3,24}$");
-
-  /**
-   * Azure Storage Container name validation. A storage container name must be between 3-63
-   * characters in length and may contain numbers, lowercase letters, and dash (-) characters only.
-   * It must start and end with a letter or number. Matching this pattern is necessary but not
-   * sufficient for a valid container name (in particular, the dash character must be immediately
-   * preceded and followed by a letter or number; consecutive dashes are not permitted).
-   */
-  public static final Pattern AZURE_STORAGE_CONTAINER_NAME_VALIDATION_PATTERN =
-      Pattern.compile("^[a-z0-9][-a-z0-9]{1,61}[a-z0-9]$");
-
-  /**
    * BigQuery datasets must be 1-1024 characters, using letters (upper or lowercase), numbers, and
    * underscores.
    */
@@ -117,13 +99,6 @@ public class ResourceValidationUtils {
   public static final Pattern RESOURCE_NAME_VALIDATION_PATTERN =
       Pattern.compile("^[a-zA-Z0-9][-_a-zA-Z0-9]{0,1023}$");
 
-  public static final Pattern AZURE_RELAY_NAMESPACE_PATTERN =
-      Pattern.compile("^[a-zA-Z0-9][a-zA-Z0-9-]{0,78}[a-zA-Z0-9]$");
-
-  /** Batch Pool id must be -64 characters, using letters, numbers, dashes, and underscores */
-  public static final Pattern AZURE_BATCH_POOL_ID_VALIDATION_PATTERN =
-      Pattern.compile("^[-_a-zA-Z0-9]{0,63}$");
-
   // An object named "." or ".." is nearly impossible for a user to delete.
   private static final ImmutableList<String> DISALLOWED_OBJECT_NAMES = ImmutableList.of(".", "..");
 
@@ -139,7 +114,6 @@ public class ResourceValidationUtils {
   private static final String GOOG_PREFIX = "goog";
   private static final ImmutableList<String> GOOGLE_NAMES = ImmutableList.of("google", "g00gle");
   private static final int MAX_RESOURCE_DESCRIPTION_NAME = 2048;
-  private static final int MAX_BATCH_POOL_DISPLAY_NAME = 1024;
   private static final int MAX_FLEXIBLE_RESOURCE_DATA_BYTE_SIZE = 5000;
 
   private final GitRepoReferencedResourceConfiguration gitRepoReferencedResourceConfiguration;
@@ -311,76 +285,6 @@ public class ResourceValidationUtils {
     }
   }
 
-  /**
-   * See
-   * https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules
-   * for azure resource rules
-   */
-  public static void validateAzureIPorSubnetName(String name) {
-    Pattern pattern = Pattern.compile("^[a-zA-Z0-9][a-zA-Z0-9-_.]{0,78}[a-zA-Z0-9_]$");
-
-    if (!pattern.matcher(name).matches()) {
-      logger.warn("Invalid Azure IP or Subnet name {}", name);
-      throw new InvalidReferenceException(
-          "Invalid Azure IP or Subnet name specified. See documentation for full specification https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules.");
-    }
-  }
-
-  /**
-   * See
-   * https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules
-   * for azure resource rules
-   */
-  public static void validateAzureNamespace(String name) {
-    if (!AZURE_RELAY_NAMESPACE_PATTERN.matcher(name).matches()
-        || name.length() > 50
-        || name.length() < 6) {
-      logger.warn("Invalid Azure Namespace {}", name);
-      throw new InvalidReferenceException(
-          "Invalid Azure Namespace specified. See documentation for full specification https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules.");
-    }
-  }
-
-  public static void validateAzureNetworkName(String name) {
-    Pattern pattern = Pattern.compile("^[a-zA-Z0-9][a-zA-Z0-9-_.]{2,62}[a-zA-Z0-9_]$");
-
-    if (!pattern.matcher(name).matches()) {
-      logger.warn("Invalid Azure network name {}", name);
-      throw new InvalidReferenceException(
-          "Invalid Azure network name specified. See documentation for full specification https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules.");
-    }
-  }
-
-  public static void validateAzureDiskName(String name) {
-    Pattern pattern = Pattern.compile("^[a-zA-Z0-9-_]{0,80}$");
-
-    if (!pattern.matcher(name).matches()) {
-      logger.warn("Invalid Disk name {}", name);
-      throw new InvalidReferenceException(
-          "Invalid Azure Disk name specified. See documentation for full specification https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules.");
-    }
-  }
-
-  public static void validateAzureBatchPoolId(String id) {
-    if (!AZURE_BATCH_POOL_ID_VALIDATION_PATTERN.matcher(id).matches()) {
-      logger.warn("Invalid Azure Batch Pool id {}", id);
-      throw new InvalidReferenceException(
-          "Invalid Azure Batch Pool id specified. Name must be 1 to 64 alphanumeric characters or underscores or dashes.");
-    }
-  }
-
-  public static void validateAzureCidrBlock(String range) {
-    Pattern pattern =
-        Pattern.compile(
-            "^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(\\/(\\d|[1-2]\\d|3[0-2]))$");
-
-    if (!pattern.matcher(range).matches()) {
-      logger.warn("Invalid CIDR block {}", range);
-      throw new InvalidReferenceException(
-          "Invalid Azure CIDR block specified. See documentation for full specification https://stackoverflow.com/questions/18608165/cidr-notation-and-ip-range-validator-pattern/18611259.");
-    }
-  }
-
   public static void validateBqDatasetName(String name) {
     if (StringUtils.isEmpty(name) || !BQ_DATASET_NAME_VALIDATION_PATTERN.matcher(name).matches()) {
       logger.warn("Invalid BQ name {}", name);
@@ -451,31 +355,6 @@ public class ResourceValidationUtils {
     }
   }
 
-  public static void validateBatchPoolDisplayName(@Nullable String displayName) {
-    if (displayName != null && displayName.length() > MAX_BATCH_POOL_DISPLAY_NAME) {
-      throw new InvalidNameException(
-          "Invalid display name specified. Display name must be under 1024 characters.");
-    }
-  }
-
-  public static void validateStorageAccountName(String storageAccountName) {
-    if (!AZURE_STORAGE_ACCOUNT_NAME_VALIDATION_PATTERN.matcher(storageAccountName).matches()) {
-      logger.warn("Invalid Storage Account name: {}", storageAccountName);
-      throw new InvalidReferenceException(
-          "Invalid Azure Storage Account name. The name must be 3 to 24 alphanumeric lower case characters.");
-    }
-  }
-
-  public static void validateStorageContainerName(String storageContainerName) {
-    if (!AZURE_STORAGE_CONTAINER_NAME_VALIDATION_PATTERN.matcher(storageContainerName).matches()
-        || storageContainerName.contains("--")) {
-      logger.warn("Invalid Storage Container name: {}", storageContainerName);
-      throw new InvalidReferenceException(
-          "Invalid Azure Storage Container name. The name must be 3 to 63 alphanumeric lower case characters "
-              + "or dashes, must start and end with a letter or number, and cannot contain consecutive dashes.");
-    }
-  }
-
   public static void validateRegion(
       TpsApiDispatch tpsApiDispatch, UUID workspaceId, String region, CloudPlatform cloudPlatform) {
     Rethrow.onInterrupted(
@@ -507,34 +386,6 @@ public class ResourceValidationUtils {
     if (fieldValue == null) {
       throw new MissingRequiredFieldException(
           String.format("Missing required field '%s' for %s", fieldName, resourceDescriptor));
-    }
-  }
-
-  public static void validateApiAzureVmCreationParameters(
-      ApiAzureVmCreationParameters apiAzureVmCreationParameters) {
-    var vmImage = apiAzureVmCreationParameters.getVmImage();
-    if (StringUtils.isEmpty(vmImage.getUri())
-        && StringUtils.isEmpty(vmImage.getPublisher())
-        && StringUtils.isEmpty(vmImage.getOffer())
-        && StringUtils.isEmpty(vmImage.getSku())
-        && StringUtils.isEmpty(vmImage.getVersion())) {
-      throw new MissingRequiredFieldException(
-          "Missing required fields for vmImage. Either uri or publisher, offer, sku, version should be defined.");
-    }
-    if (StringUtils.isEmpty(vmImage.getUri())
-        && (StringUtils.isEmpty(vmImage.getPublisher())
-            || StringUtils.isEmpty(vmImage.getOffer())
-            || StringUtils.isEmpty(vmImage.getSku())
-            || StringUtils.isEmpty(vmImage.getVersion()))) {
-      throw new MissingRequiredFieldException(
-          "Missing required fields for vmImage. Publisher, offer, sku, version should be defined.");
-    }
-    if (StringUtils.isEmpty(vmImage.getUri())
-        && (!StringUtils.isEmpty(vmImage.getPublisher())
-            && !StringUtils.isEmpty(vmImage.getOffer())
-            && !StringUtils.isEmpty(vmImage.getSku())
-            && !StringUtils.isEmpty(vmImage.getVersion()))) {
-      checkFieldNonNull(apiAzureVmCreationParameters.getVmUser(), "vmUser");
     }
   }
